Move pools constants back into config
