+++ balancer-js/src/lib/constants/config.ts
@@ -20,72 +20,11 @@ export const BALANCER_NETWORK_CONFIG: Record<Network, BalancerNetworkConfig> = {
             subgraph:
                 'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-v2',
         },
-    },
-    [Network.ROPSTEN]: {
-        chainId: Network.ROPSTEN, //3
-        addresses: {
-            contracts: {
-                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
-                multicall: '0x53c43764255c17bd724f74c4ef150724ac50a3ed',
+        pools: {
+            staBal3Pool: {
+                id: '0x7b50775383d3d6f0215a8f290f2c9e2eebbeceb20000000000000000000000fe',
+                address: '0x7b50775383d3d6f0215a8f290f2c9e2eebbeceb2',
             },
-            tokens: {
-                wrappedNativeAsset:
-                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
-            },
-        },
-        urls: {
-            subgraph: '',
-        },
-    },
-    [Network.RINKEBY]: {
-        chainId: Network.RINKEBY, //4
-        addresses: {
-            contracts: {
-                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
-                multicall: '0x42ad527de7d4e9d9d011ac45b31d8551f8fe9821',
-            },
-            tokens: {
-                wrappedNativeAsset:
-                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
-            },
-        },
-        urls: {
-            subgraph:
-                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-rinkeby-v2',
-        },
-    },
-    [Network.GÖRLI]: {
-        chainId: Network.GÖRLI, //5
-        addresses: {
-            contracts: {
-                vault: '0x65748E8287Ce4B9E6D83EE853431958851550311',
-                multicall: '0x42ad527de7d4e9d9d011ac45b31d8551f8fe9821',
-            },
-            tokens: {
-                wrappedNativeAsset:
-                    '0x9A1000D492d40bfccbc03f413A48F5B6516Ec0Fd',
-            },
-        },
-        urls: {
-            subgraph:
-                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-goerli-v2',
-        },
-    },
-    [Network.KOVAN]: {
-        chainId: Network.KOVAN, //42
-        addresses: {
-            contracts: {
-                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
-                multicall: '0x2cc8688C5f75E365aaEEb4ea8D6a480405A48D2A',
-            },
-            tokens: {
-                wrappedNativeAsset:
-                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
-            },
-        },
-        urls: {
-            subgraph:
-                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-kovan-v2',
         },
     },
     [Network.POLYGON]: {
@@ -104,6 +43,7 @@ export const BALANCER_NETWORK_CONFIG: Record<Network, BalancerNetworkConfig> = {
             subgraph:
                 'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-polygon-v2',
         },
+        pools: {},
     },
     [Network.ARBITRUM]: {
         chainId: Network.ARBITRUM, //42161
@@ -121,5 +61,86 @@ export const BALANCER_NETWORK_CONFIG: Record<Network, BalancerNetworkConfig> = {
             subgraph:
                 'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-arbitrum-v2',
         },
+        pools: {},
+    },
+    [Network.KOVAN]: {
+        chainId: Network.KOVAN, //42
+        addresses: {
+            contracts: {
+                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+                multicall: '0x2cc8688C5f75E365aaEEb4ea8D6a480405A48D2A',
+            },
+            tokens: {
+                wrappedNativeAsset:
+                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
+            },
+        },
+        urls: {
+            subgraph:
+                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-kovan-v2',
+        },
+        pools: {
+            staBal3Pool: {
+                id: '0x8fd162f338b770f7e879030830cde9173367f3010000000000000000000004d8',
+                address: '0x8fd162f338b770f7e879030830cde9173367f301',
+            },
+            wethStaBal3: {
+                id: '0x6be79a54f119dbf9e8ebd9ded8c5bd49205bc62d00020000000000000000033c',
+                address: '0x6be79a54f119dbf9e8ebd9ded8c5bd49205bc62d',
+            },
+        },
+    },
+    [Network.ROPSTEN]: {
+        chainId: Network.ROPSTEN, //3
+        addresses: {
+            contracts: {
+                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+                multicall: '0x53c43764255c17bd724f74c4ef150724ac50a3ed',
+            },
+            tokens: {
+                wrappedNativeAsset:
+                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
+            },
+        },
+        urls: {
+            subgraph: '',
+        },
+        pools: {},
+    },
+    [Network.RINKEBY]: {
+        chainId: Network.RINKEBY, //4
+        addresses: {
+            contracts: {
+                vault: '0xBA12222222228d8Ba445958a75a0704d566BF2C8',
+                multicall: '0x42ad527de7d4e9d9d011ac45b31d8551f8fe9821',
+            },
+            tokens: {
+                wrappedNativeAsset:
+                    '0xdFCeA9088c8A88A76FF74892C1457C17dfeef9C1',
+            },
+        },
+        urls: {
+            subgraph:
+                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-rinkeby-v2',
+        },
+        pools: {},
+    },
+    [Network.GÖRLI]: {
+        chainId: Network.GÖRLI, //5
+        addresses: {
+            contracts: {
+                vault: '0x65748E8287Ce4B9E6D83EE853431958851550311',
+                multicall: '0x42ad527de7d4e9d9d011ac45b31d8551f8fe9821',
+            },
+            tokens: {
+                wrappedNativeAsset:
+                    '0x9A1000D492d40bfccbc03f413A48F5B6516Ec0Fd',
+            },
+        },
+        urls: {
+            subgraph:
+                'https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-goerli-v2',
+        },
+        pools: {},
     },
 };
